# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: contacts
    description: Contact/lead form submissions
  - name: openrouter
    description: OpenRouter AI chat operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /contacts:
    post:
      operationId: createContact
      tags: [contacts]
      summary: Submit a contact/lead form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "201":
          description: Contact submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all contact submissions
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
  /openrouter/conversations:
    get:
      operationId: listOpenrouterConversations
      tags: [openrouter]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenrouterConversation"
    post:
      operationId: createOpenrouterConversation
      tags: [openrouter]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenrouterConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterConversation"
  /openrouter/conversations/{id}:
    get:
      operationId: getOpenrouterConversation
      tags: [openrouter]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterError"
    delete:
      operationId: deleteOpenrouterConversation
      tags: [openrouter]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterError"
  /openrouter/conversations/{id}/messages:
    get:
      operationId: listOpenrouterMessages
      tags: [openrouter]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenrouterMessage"
    post:
      operationId: sendOpenrouterMessage
      tags: [openrouter]
      summary: Send a message and receive an AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenrouterMessageBody"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CreateContactBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        plan:
          type: string
        message:
          type: ["string", "null"]
      required:
        - name
        - phone
        - email
        - plan
    Contact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        plan:
          type: string
        message:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - phone
        - email
        - plan
        - status
        - createdAt
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    OpenrouterConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    OpenrouterMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateOpenrouterConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendOpenrouterMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    OpenrouterConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenrouterMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    OpenrouterError:
      type: object
      properties:
        error:
          type: string
      required:
        - error