# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Submit a contact/lead form
 */
export const CreateContactBody = zod.object({
  name: zod.string(),
  phone: zod.string(),
  email: zod.string(),
  plan: zod.string(),
  message: zod.string().nullish(),
});

/**
 * @summary List all contact submissions
 */
export const ListContactsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string(),
  email: zod.string(),
  plan: zod.string(),
  message: zod.string().nullish(),
  status: zod.string(),
  createdAt: zod.string(),
});
export const ListContactsResponse = zod.array(ListContactsResponseItem);

/**
 * @summary List all conversations
 */
export const ListOpenrouterConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenrouterConversationsResponse = zod.array(
  ListOpenrouterConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateOpenrouterConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetOpenrouterConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenrouterConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteOpenrouterConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListOpenrouterMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenrouterMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenrouterMessagesResponse = zod.array(
  ListOpenrouterMessagesResponseItem,
);

/**
 * @summary Send a message and receive an AI response (SSE stream)
 */
export const SendOpenrouterMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenrouterMessageBody = zod.object({
  content: zod.string(),
});